chore: organize summary dashboard code
