+++ src/scenes/Summary/index.ts
@@ -1,0 +1,1 @@
+export { getSummaryScene } from './summaryScene';
